Fjerner default survey
