+++ demo/src/index.tsx
@@ -18,6 +18,13 @@ class Demo extends Component {
                     queueKey='Q_CHAT_BOT'
                     configId='599f9e7c-7f6b-4569-81a1-27202c419953'
                     label={'Chat med NAV'}
+                    analyticsSurvey={[
+                        {
+                            label: 'Fikk du svar på det du lurte på?',
+                            options: ['Ja', 'Nei', 'Delvis'],
+                            type: 'radio',
+                        },
+                    ]}
                 />
             </Outer>
         );
