+++ src/index.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 import { ThemeProvider } from 'styled-components';
 import ChatContainer from './components/ChatContainer';
 import tema from './tema/tema';
-import { SurveyQuestion } from './components/Evaluering/surveyFields';
+import { SurveyQuestion } from './components/Evaluering';
 
 export type AnalyticsCallback = (event: string, data: any) => void;
